Update the simplifier to respect and apply the 'use var' user preference.   (#25865)
* Initial UseConditional work.

* Initial UseConditional work.

* Working replacement.

* Extract common code.

* Make 'var' work properly in the simplifier service.

* Move types to their own files.

* Provide a way to create types that are not simplified to 'var'.

* Update test options.

* Update test.

* Move CSharp code style code down to the Workspace layer.

* Rename config value.

* Add ui options.

* Add VB ui options.

* Add VB support

* Strip trivia

* Initial stubs for 'use conditional for returns'

* Add VB impl.

* Only support processing of real trees.

* Actually assign variable.

* Don't assume a single variable.

* Share more code.

* Fix shared code.

* Add check.

* Simplify code.

* Only the analyzer needs to check the severity.

* Generalize loc strings.

* remove static.

* Remove unneeded methods.

* Do cheap checks first.

* Handle more cases.

* Handle more cases.

* Add test.

* Add tests.

* Add test.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Fix tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Add tests.

* Format multiline conditions.

* Add multiline formatting for VB.

* Move code.

* Simpler formilization of casts.

* Extract out code.

* Remove file.

* Move files.

* Share more code.

* Share more code.

* Break out types.

* Share more code.

* Add more tests.

* Add VB tests.

* Simplify.

* Add comment.

* Remove newlines.

* Move method.

* Move method.

* Add comments.

* Simplify

* Update comment.

* Update comment.

* Remove line.

* Update comment.

* Update comment.

* Update comment.

* Fix comment.

* Wrap long lines.

* Wrap long lines.

* Wrap long lines.

* Handle trivia.

* Simplify.

* Fix check.

* better support for else-if chains.

* Add comments.

* Remove 'Use Conditional Expressoin'

* Remove more vestiges of
 feature.

* Remove code.

* Add back newline.
