+++ src/Workspaces/CSharp/Portable/CodeStyle/TypeStyle/TypeStyleHelper.cs
@@ -14,9 +14,9 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle
     internal static class TypeStyleHelper
     {
         /// <summary>
-        /// Given an expression of assignment, answers whether the declaration
-        /// can use var keyword, by looking at the user's style preferences 
-        /// obtained from options and the context obtained from the expression.
+        /// Given an expression or assignment, answers whether the declaration can use var keyword,
+        /// by looking at the user's style preferences obtained from options and the context
+        /// obtained from the expression.
         /// </summary>
         public static bool IsImplicitTypePreferred(
             ExpressionSyntax initializerExpression,
@@ -28,10 +28,8 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle
 
             var isBuiltInTypeContext = IsBuiltInType(initializerExpression, semanticModel, cancellationToken);
 
-            var isTypeApparentContext = IsTypeApparentInAssignmentExpression(stylePreferences,
-                                            initializerExpression,
-                                            semanticModel,
-                                            cancellationToken);
+            var isTypeApparentContext = IsTypeApparentInAssignmentExpression(
+                stylePreferences, initializerExpression, semanticModel, cancellationToken);
 
             return IsImplicitStylePreferred(stylePreferences, isBuiltInTypeContext, isTypeApparentContext);
         }
@@ -130,7 +128,7 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle
             // other Conversion cases:
             //      a. conversion with helpers like: int.Parse methods
             //      b. types that implement IConvertible and then invoking .ToType()
-            //      c. System.Convert.Totype()
+            //      c. System.Convert.ToType()
             var memberName = GetRightmostInvocationExpression(initializerExpression).GetRightmostName();
             if (memberName == null)
             {
@@ -306,7 +304,7 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle
             }
             else
             {
-                return (TypeSyntax)generator.TypeExpression(type);
+                return type.GenerateTypeSyntax(allowVar: false);
             }
         }
     }
