+++ src/Workspaces/Core/Portable/LanguageServices/SyntaxFactsService/ISyntaxFactsService.cs
@@ -348,6 +348,8 @@ namespace Microsoft.CodeAnalysis.LanguageServices
         Location GetDeconstructionReferenceLocation(SyntaxNode node);
 
         SyntaxToken? GetDeclarationIdentifierIfOverride(SyntaxToken token);
+
+        bool SpansPreprocessorDirective(IEnumerable<SyntaxNode> nodes);
     }
 
     [Flags]
